Workspace update.

+++ firmware_applications/Generic_Push_Button/STM32CubeIDE/Core/Src/main.c
@@ -312,19 +312,19 @@ void mcuConfig(void) {
 	/*			THIS VALUE CAN BE FOUND IN CREDENTIALS 						*/
 
 	ST_RF_API_set_freq_offset(sfxConfiguration.freqOffset);
-	printf("Freq Offset %d \n", sfxConfiguration.freqOffset);
+	printf("Freq Offset %d \n", (int32_t)sfxConfiguration.freqOffset);
 
 	/*			SET LBT OFFSET																		*/
 	/*			THIS VALUE CAN BE FOUND IN CREDENTIALS 						*/
 
 	ST_RF_API_set_lbt_thr_offset(sfxConfiguration.lbtOffset);
-	printf("LBT %d \n", sfxConfiguration.lbtOffset);
+	printf("LBT %d \n", (int32_t)sfxConfiguration.lbtOffset);
 
 	/*			SET RSSI OFFSET																		*/
 	/*			THIS VALUE CAN BE FOUND IN CREDENTIALS 						*/
 
 	ST_RF_API_set_rssi_offset(sfxConfiguration.rssiOffset);
-	printf("RSSI %d \n", sfxConfiguration.rssiOffset);
+	printf("RSSI %d \n", (int32_t)sfxConfiguration.rssiOffset);
 }
 
 void ST_Init(void)
